apps/api/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# AI & ML
google-generativeai>=0.3.0
pinecone-client==3.0.0
elasticsearch[async]>=8.0.0
motor>=3.3.0  # Async MongoDB driver

# Environment Management
python-dotenv==1.0.0
anyio>=3.0.0

# Optional: Development tools
# pytest==7.4.3
# httpx==0.26.0  # For testing FastAPI
# black==23.12.1  # Code formatting
# mypy==1.8.0  # Type checking
